+++ actix-web-httpauth/src/headers/authorization/scheme/mod.rs
@@ -7,9 +7,8 @@ pub mod bearer;
 
 use crate::headers::authorization::errors::ParseError;
 
-/// Authentication scheme for [`Authorization`](./struct.Authorization.html)
-/// header.
+/// Authentication scheme for [`Authorization`](super::Authorization) header.
 pub trait Scheme: TryIntoHeaderValue + Debug + Display + Clone + Send + Sync {
-    /// Try to parse the authentication scheme from the `Authorization` header.
+    /// Try to parse an authentication scheme from the `Authorization` header.
     fn parse(header: &HeaderValue) -> Result<Self, ParseError>;
 }