+++ actix-web-httpauth/src/headers/www_authenticate/mod.rs
@@ -1,9 +1,7 @@
-//! `WWW-Authenticate` header and various auth challenges
+//! `WWW-Authenticate` header and various auth challenges.
 
 mod challenge;
 mod header;
 
-pub use self::challenge::basic;
-pub use self::challenge::bearer;
-pub use self::challenge::Challenge;
+pub use self::challenge::{basic, bearer, Challenge};
 pub use self::header::WwwAuthenticate;